Rename and move to separate package controllers for admin panel endpoints
