+++ src/main/java/pl/coderslab/charity/controller/admin/DashboardController.java
@@ -1,4 +1,4 @@
-package pl.coderslab.charity.controller;
+package pl.coderslab.charity.controller.admin;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequestMapping("admin")
-class AdminDashboardController {
+class DashboardController {
 
     @GetMapping
     public String panel() {
